+++ test/util/mvc/controllers/gracefullShutDownController.js
@@ -1,8 +1,8 @@
-module.exports = (router) => {
+module.exports = router => {
   router.route('/longWait').get((req, res) => {
     process.send('waiting')
     setTimeout(() => {
       res.send('longWait done')
-    }, 10000)
+    }, 1000)
   })
 }